refactor : endGame() 분리
스트라이크가 3개일 경우 생성되는 endGame()을 endGame.js로 분리

#issue 1 : endGame() 호출 시 setAnswer()가 2번 호출되는 것 같음
정답이 2번 초기화 됨
