+++ src/result-const.js
